Убрал пробелы между оператором цикла и скобкой.
